Updated Javadocs throughout entire package
git-svn-id: https://svn.cloudhopper.lan/repos/dev/utility/java/ch-commons-xbean/trunk@674 db02c596-f934-4852-9b2a-e10ddaadd22d

+++ src/main/java/com/cloudhopper/commons/xbean/convert/ShortPropertyConverter.java
@@ -4,13 +4,14 @@ package com.cloudhopper.commons.xbean.convert;
 import com.cloudhopper.commons.xbean.*;
 
 /**
+ * Converts a String to a Short.
  * @author joelauer
  */
 public class ShortPropertyConverter implements PropertyConverter {
 
     public Object convert(String value) throws ConversionException {
         try {
-            return Byte.valueOf(value);
+            return Short.valueOf(value);
         } catch (NumberFormatException e) {
             throw new ConversionException(e.getMessage());
         }
